+++ back/wordseed/src/main/java/com/spring/wordseed/controller/PostController.java
@@ -3,6 +3,7 @@ package com.spring.wordseed.controller;
 import com.spring.wordseed.dto.in.*;
 import com.spring.wordseed.dto.out.*;
 import com.spring.wordseed.enu.*;
+import com.spring.wordseed.service.CommentService;
 import com.spring.wordseed.service.PostService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -11,15 +12,16 @@ import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.List;
 
 @RestController
 @RequestMapping("/post")
 public class PostController {
     private final PostService postService;
+    private final CommentService commentService;
     @Autowired
-    PostController(PostService postService){
+    PostController(PostService postService, CommentService commentService){
         this.postService = postService;
+        this.commentService = commentService;
     }
     // 작품 업로드
     @PostMapping("")
@@ -105,7 +107,7 @@ public class PostController {
     @PostMapping("/comment")
     public ResponseEntity<CreateCommentOutDTO> createComment(@RequestBody CreateCommentInDTO createCommentInDTO) throws Exception {
         // add request for userId
-        CreateCommentOutDTO createCommentOutDTO = postService.createComment(createCommentInDTO);
+        CreateCommentOutDTO createCommentOutDTO = commentService.createComment(createCommentInDTO);
 
         return ResponseEntity.status(HttpStatus.OK).body(createCommentOutDTO);
     }
