+++ back/wordseed/src/main/java/com/spring/wordseed/service/CommentService.java
@@ -1,0 +1,8 @@
+package com.spring.wordseed.service;
+
+import com.spring.wordseed.dto.in.CreateCommentInDTO;
+import com.spring.wordseed.dto.out.CreateCommentOutDTO;
+
+public interface CommentService {
+    CreateCommentOutDTO createComment(CreateCommentInDTO createCommentInDTO) throws Exception;
+}
