+++ back/wordseed/src/main/java/com/spring/wordseed/service/PostService.java
@@ -13,7 +13,6 @@ public interface PostService {
     ReadPostByPostIdOutDTO readPostByPostId(Long postId);
     UpdatePostOutDTO updatePost(UpdatePostInDTO updatePostInDTO) throws Exception;
     void deletePost(DeletePostInDTO deletePostInDTO);
-    CreateCommentOutDTO createComment(CreateCommentInDTO createCommentInDTO) throws Exception;
     UpdateCommentOutDTO UpdateComment(UpdateCommentInDTO updateCommentInDTO);
     void deleteComment(DeleteCommentInDTO deleteCommentInDTO);
     CreateLikeOutDTO createLike(CreateLikeInDTO createLikeInDTO);
