+++ back/wordseed/src/main/java/com/spring/wordseed/service/impl/PostServiceImpl.java
@@ -131,25 +131,7 @@ public class PostServiceImpl implements PostService {
 
     }
 
-    @Override
-    public CreateCommentOutDTO createComment(CreateCommentInDTO createCommentInDTO) throws Exception {
-        Comment comment = Comment.builder()
-                .content(createCommentInDTO.getContent())
-                .user(userRepo.findById(100L).orElseThrow(Exception::new)) // error
-                .post(postRepo.findById(createCommentInDTO.getPostId()).orElseThrow(Exception::new))
-                .build();
 
-        commentRepo.save(comment);
-
-        return CreateCommentOutDTO.builder()
-                .commentId(comment.getCommentId())
-                .userId(comment.getUser().getUserId())
-                .postId(comment.getPost().getPostId())
-                .content(comment.getContent())
-                .createdAt(comment.getCreatedAt())
-                .updatedAt(comment.getUpdatedAt())
-                .build();
-    }
 
     @Override
     public UpdateCommentOutDTO UpdateComment(UpdateCommentInDTO updateCommentInDTO) {
